Relocate modules to package __init__
